+++ demos/manual/src/main/java/is/codion/manual/notes/NotesDemo.java
@@ -190,7 +190,7 @@ public final class NotesDemo {
 							// Note.CREATED is excluded by default since it is not updatable.
 							.editable(attributes -> attributes.remove(Note.UPDATED)));
 			// Configure the table and columns
-			table().model().sorter().descending(Note.CREATED);
+			table().model().sort().descending(Note.CREATED);
 			table().setAutoResizeMode(JTable.AUTO_RESIZE_SUBSEQUENT_COLUMNS);
 			FilterTableColumnModel<Attribute<?>> columnModel = table().columnModel();
 			columnModel.column(Note.NOTE).setPreferredWidth(280);