+++ swing/common-model/src/test/java/is/codion/swing/common/model/component/table/DefaultFilterTableModelTest.java
@@ -230,7 +230,7 @@ public final class DefaultFilterTableModelTest {
 										.build();
 		testModel.selection().indexes().addListener(selectionEvents::incrementAndGet);
 		testModel.items().refreshStrategy().set(RefreshStrategy.MERGE);
-		testModel.sorter().ascending(0);
+		testModel.sort().ascending(0);
 		testModel.items().refresh();
 		testModel.selection().index().set(1);//b
 
@@ -266,7 +266,7 @@ public final class DefaultFilterTableModelTest {
 
 	@Test
 	void addItems() {
-		tableModel.sorter().ascending(0);
+		tableModel.sort().ascending(0);
 		Items<TestRow> items = tableModel.items();
 		items.add(asList(A, B));
 		// sorts
@@ -748,7 +748,7 @@ public final class DefaultFilterTableModelTest {
 
 	@Test
 	void replace() {
-		tableModel.sorter().ascending(0);
+		tableModel.sort().ascending(0);
 		tableModel.items().refresh();
 		//a, b, c, d, e
 		//replace d with f
@@ -834,7 +834,7 @@ public final class DefaultFilterTableModelTest {
 						.build();
 
 		model.items().refresh();
-		model.sorter().ascending(0);
+		model.sort().ascending(0);
 
 		assertTrue(model.items().contains(A));
 		assertFalse(model.items().contains(G));