+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/FilterTableHeaderRenderer.java
@@ -19,7 +19,7 @@
 package is.codion.swing.common.ui.component.table;
 
 import is.codion.common.model.condition.ConditionModel;
-import is.codion.swing.common.model.component.table.FilterTableSorter.ColumnSortOrder;
+import is.codion.swing.common.model.component.table.FilterTableSort.ColumnSortOrder;
 
 import javax.swing.AbstractButton;
 import javax.swing.Icon;
@@ -80,7 +80,7 @@ final class FilterTableHeaderRenderer<R, C> implements TableCellRenderer {
 	}
 
 	private Icon sortArrowIcon(C identifier, int iconSizePixels) {
-		ColumnSortOrder<C> columnSortOrder = filterTable.model().sorter().columns().get(identifier);
+		ColumnSortOrder<C> columnSortOrder = filterTable.model().sort().columns().get(identifier);
 
 		return columnSortOrder.sortOrder() == SortOrder.UNSORTED ? null : new Arrow(columnSortOrder.sortOrder() == SortOrder.DESCENDING, iconSizePixels, columnSortOrder.priority());
 	}