+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -912,7 +912,7 @@ public final class EntitySearchField extends HintTextField {
 											.enable(t.searchField()))
 							.build();
 
-			filterTable.model().sorter().ascending(searchField.model.columns().iterator().next());
+			filterTable.model().sort().ascending(searchField.model.columns().iterator().next());
 			filterTable.columnModel().visible().set(searchField.model.columns().toArray(new Attribute[0]));
 
 			return filterTable;