+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.34
 ### is.codion.common.model
 - FilterModel.VisibleItems and FilteredItems.get() annotated as NonNull.
+- FilterModel.Sorter renamed Sort, sorter() renamed sort().
 ### is.codion.framework.domain
 - Column.Getter and Setter renamed GetValue and SetParameter respectively.
 ### is.codion.swing.common.ui